Rename ApiErrorPlaceholder to ErrorPlaceholder
