+++ assets/js/blocks/reviews/reviews-by-product/no-reviews-placeholder.js
@@ -8,13 +8,13 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import ApiErrorPlaceholder from '../../../components/api-error-placeholder';
+import ErrorPlaceholder from '../../../components/error-placeholder';
 import { IconReviewsByProduct } from '../../../components/icons';
 import { withProduct } from '../../../hocs';
 
 const NoReviewsPlaceholder = ( { error, getProduct, isLoading, product } ) => {
 	const renderApiError = () => (
-		<ApiErrorPlaceholder
+		<ErrorPlaceholder
 			className="wc-block-featured-product-error"
 			error={ error }
 			isLoading={ isLoading }
