+++ assets/js/components/error-placeholder/index.js
@@ -14,7 +14,7 @@ import { Button, Placeholder, Spinner } from '@wordpress/components';
 import ErrorMessage from './error-message.js';
 import './style.scss';
 
-const ApiErrorPlaceholder = ( { className, error, isLoading, onRetry } ) => (
+const ErrorPlaceholder = ( { className, error, isLoading, onRetry } ) => (
 	<Placeholder
 		icon={ <Gridicon icon="notice" /> }
 		label={ __(
@@ -38,7 +38,7 @@ const ApiErrorPlaceholder = ( { className, error, isLoading, onRetry } ) => (
 	</Placeholder>
 );
 
-ApiErrorPlaceholder.propTypes = {
+ErrorPlaceholder.propTypes = {
 	/**
 	 * Classname to add to placeholder in addition to the defaults.
 	 */
@@ -67,4 +67,4 @@ ApiErrorPlaceholder.propTypes = {
 	onRetry: PropTypes.func,
 };
 
-export default ApiErrorPlaceholder;
+export default ErrorPlaceholder;
